+++ Sushi/src/main/java/com/zetterlund/wigell_sushi_api/repository/BookingRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface BookingRepository extends JpaRepository<Booking, Long> {
-    List<Booking> findByCustomerId(Long customerId);
+public interface BookingRepository extends JpaRepository<Booking, Integer> {
+    List<Booking> findByCustomerId(Integer customerId);
 }
